fix(parity): align bonus timer float32 updates

+++ src/crimson/bonuses/double_experience.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from ..math_parity import f32
 from .apply_context import BonusApplyCtx, bonus_apply_seconds
 
 
@@ -7,4 +8,6 @@ def apply_double_experience(ctx: BonusApplyCtx) -> None:
     old = float(ctx.state.bonuses.double_experience)
     if old <= 0.0:
         ctx.register_global("double_experience")
-    ctx.state.bonuses.double_experience = float(old + bonus_apply_seconds(ctx) * ctx.economist_multiplier)
+    ctx.state.bonuses.double_experience = float(
+        f32(float(old) + bonus_apply_seconds(ctx) * float(ctx.economist_multiplier))
+    )
