+++ src/crimson/bonuses/energizer.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from ..math_parity import f32
 from .apply_context import BonusApplyCtx, bonus_apply_seconds
 
 
@@ -8,4 +9,6 @@ def apply_energizer(ctx: BonusApplyCtx) -> None:
     if old <= 0.0:
         ctx.register_global("energizer")
 
-    ctx.state.bonuses.energizer = float(old + bonus_apply_seconds(ctx) * ctx.economist_multiplier)
+    ctx.state.bonuses.energizer = float(
+        f32(float(old) + bonus_apply_seconds(ctx) * float(ctx.economist_multiplier))
+    )
